Move test_utils to its own file
